fix tilling dtype

+++ stlearn/image_preprocessing/image_tiling.py
@@ -52,11 +52,7 @@ def tiling(
     if not os.path.isdir(out_path):
         os.mkdir(out_path)
 
-    image = (
-        adata.uns["spatial"][library_id]["images"][adata.uns["spatial"]["use_quality"]]
-        * 255
-    ).astype(np.uint8)
-
+    image = adata.uns["spatial"][library_id]["images"][adata.uns["spatial"]["use_quality"]]
     img_pillow = Image.fromarray(image)
     tile_names = []
